safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/change-iframe-src-expected.txt

CONSOLE MESSAGE: line 6: IFRAME HOST: 127.0.0.1:8000
CONSOLE MESSAGE: line 6: IFRAME HOST: localhost:8000

Tests that Elements panel allows to change src attribute on iframes inside inspected page. See bug 41350.
